css, jsp 수정

+++ The_First_modify/src/job_community/action/JobBoardDeleteProAction.java
@@ -1,12 +1,15 @@
 package job_community.action;
 
 import java.io.PrintWriter;
+import java.util.ArrayList;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import job_community.svc.JobBoardDeleteProService;
 import job_community.svc.JobBoardDetailService;
+import job_community.svc.JobCommentDeleteProService;
+import job_community.svc.JobCommentListService;
 import job_community.vo.ActionForward;
 import job_community.vo.JobBoardBean;
 import member.svc.MemberUpdateProService;
@@ -24,26 +27,30 @@ public class JobBoardDeleteProAction implements Action {
 		JobBoardDetailService jobBoardDetailService = new JobBoardDetailService();
 		JobBoardBean jobBoardBean = jobBoardDetailService.getArticle(num);
 		String nickname = jobBoardBean.getNickname();
-		 
-			boolean isDeleteSuccess = boardDeleteProService.removeArticle(num);
-					
-			if(!isDeleteSuccess) {
-				response.setContentType("text/html; charset=UTF-8");
-				PrintWriter out = response.getWriter();
-				out.println("<script>");
-				out.println("alert('글 삭제 실패!')");
-				out.println("history.back()");
-				out.println("</script>");
-			} else {
-				MemberUpdateProService memberUpdateProService = new MemberUpdateProService();
-				boolean isSuccess = memberUpdateProService.minusArticletLP(nickname);
-				if(isSuccess) {
-					forward = new ActionForward();
-					forward.setPath("JobBoardList.job");
-					forward.setRedirect(true);
+		
+		JobCommentDeleteProService jobCommentDeleteProService = new JobCommentDeleteProService();
+		boolean isDeleteSuccess  = jobCommentDeleteProService.deleteCommentAll(num);
+		
+		if(isDeleteSuccess) {
+			isDeleteSuccess = boardDeleteProService.removeArticle(num);
+				if(!isDeleteSuccess) {
+					response.setContentType("text/html; charset=UTF-8");
+					PrintWriter out = response.getWriter();
+					out.println("<script>");
+					out.println("alert('글 삭제 실패!')");
+					out.println("history.back()");
+					out.println("</script>");
+				} else {
+					MemberUpdateProService memberUpdateProService = new MemberUpdateProService();
+					boolean isSuccess = memberUpdateProService.minusArticletLP(nickname);
+					if(isSuccess) {
+						forward = new ActionForward();
+						forward.setPath("JobBoardList.job");
+						forward.setRedirect(true);
+					}
 				}
-			}
-					
+		}
+			
 		return forward;
 	}
